+++ packages/calcite-components/src/components/color-picker-hex-input/color-picker-hex-input.e2e.ts
@@ -2,7 +2,7 @@
 import { newE2EPage, E2EPage, E2EElement } from "@arcgis/lumina-compiler/puppeteerTesting";
 import { describe, expect, it, beforeEach } from "vitest";
 import { accessible, defaults, focusable, hidden, reflects, renders } from "../../tests/commonTests";
-import { selectText } from "../../tests/utils";
+import { selectText } from "../../tests/utils/puppeteer";
 import { canConvertToHexa, isValidHex, normalizeHex } from "../color-picker/utils";
 import { CSS } from "./resources";